fix: widen _buffer_size/_max_time_lost uint8_t->int (lift ~10s lost-track window cap)
_buffer_size and _max_time_lost were declared uint8_t, which caps at 255.
At 25 fps and track_buffer=500 the intended window is ~416 frames (~16 s),
but uint8_t silently wraps to ~160 — cutting the re-id gallery window to ~6 s.

Split the declaration so the two small config inputs (_track_buffer, _frame_rate)
stay uint8_t while the computed window members become int.  The construction
cast is updated to match.  All four use-sites are arithmetic/comparison; no
format specifiers or serialisation involved — value-preserving change.

Part of RC-1 match-duration re-id gallery work.

+++ botsort/src/BoTSORT.cpp
@@ -61,7 +61,7 @@ BoTSORT::BoTSORT(const Config<TrackerParams> &tracker_config,
 
     // Tracker module
     _frame_id = 0;
-    _buffer_size = static_cast<uint8_t>(_frame_rate / 30.0 * _track_buffer);
+    _buffer_size = static_cast<int>(_frame_rate / 30.0 * _track_buffer);
     _max_time_lost = _buffer_size;
     _kalman_filter = std::make_unique<KalmanFilter>(
             static_cast<double>(1.0 / _frame_rate));
